requirements.txt
# RedTeam-Agent — Author: Abdulwahab Hamoud Salah

python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
sublist3r==1.0
python-nmap==0.7.1
python-whois==0.8.0
reportlab==4.0.7
rich==13.7.0
anthropic>=0.40.0
openai>=1.30.0
dnspython==2.4.2
colorama==0.4.6
fpdf2==2.7.6
tqdm==4.66.1
